feat: new module for modals
