+++ js/util.js
@@ -13,34 +13,4 @@ function isStringFits(string, maxLength) {
   return typeof string === 'string' ? string.length <= maxLength : null;
 }
 
-function closeModal(evt) {
-  if (evt.code === 'Escape' || evt.type === 'click') {
-    this.window.classList.add('hidden');
-    document.body.classList.remove('modal-open');
-
-    this.button.removeEventListener('click', this);
-    document.removeEventListener('keydown', this);
-  }
-}
-
-function getCloseButton (modal) {
-  return (modal.querySelector('.cancel') || modal.querySelector('.success__button') || modal.querySelector('.error__button'));
-}
-
-function openModal(modalSelector) {
-  const currentModal = document.querySelector(modalSelector);
-  const closeButton = getCloseButton(currentModal);
-  const objListenerCallback = {
-    handleEvent: closeModal,
-    window: currentModal,
-    button: closeButton,
-  };
-
-  closeButton.addEventListener('click', objListenerCallback);
-  document.addEventListener('keydown', objListenerCallback);
-
-  currentModal.classList.remove('hidden');
-  document.body.classList.add('modal-open');
-}
-
-export {getRandomInt, isStringFits, openModal};
+export {getRandomInt, isStringFits};
